+++ lib/bmg/support.rb
@@ -1,3 +1,4 @@
 require_relative 'support/tuple_algebra'
 require_relative 'support/tuple_transformer'
 require_relative 'support/keys'
+require_relative 'support/output_preferences'